✨ Preparation Single File Components
- [x] create `App.vue`,  and change `popup.js`
- [x] install style-loader nad css-loader
