+++ webpack.config.js
@@ -39,6 +39,10 @@ module.exports = function(webpackEnv, options) {
         {
           test: /\.vue$/,
           loader: 'vue-loader'
+        },
+        {
+          test: /\.css$/,
+          use: ['style-loader', 'css-loader']
         }
       ]
     },
